requirements.txt
streamlit==1.32.0
torch==2.1.0
torchvision==0.16.0
Pillow==10.2.0
numpy==1.26.3
opencv-python==4.9.0.80
